+++ tests/integration/test_scripting_engine_per.gd
@@ -122,3 +122,24 @@ func test_per_boardseek():
 	yield(yield_for(0.3), YIELD)
 	assert_eq(hand.get_card_count(), 3,
 			"Draw 1 card per 0-cost card on board")
+
+
+func test_per_counter():
+	board.counters.mod_counter("research", 3)
+	card.scripts = {"manual": {
+		"hand": [
+			{"name": "move_card_to_container",
+			"subject": "index",
+			"subject_count": "per_counter",
+			"src_container": deck,
+			"dest_container": hand,
+			"subject_index": "top",
+			"per_counter": {
+				"counter_name": "research"}
+			},
+		]}
+	}
+	card.execute_scripts()
+	yield(yield_for(0.5), YIELD)
+	assert_eq(hand.get_card_count(), 8,
+		"Draw 1 card per counter specified")